feat(fonts): font changes will be stored in localStorage
