+++ src/components/MenuBar/MenuItems/ChangeFontBody.tsx
@@ -1,26 +1,6 @@
 import React, { useState, MutableRefObject } from "react";
 import { changeFont } from "../../../helpers";
-
-const fontFamilies: string[] = [
-  "Montserrat",
-  "Nanum Gothic",
-  "Nunito",
-  "Oswald",
-  "Poppins",
-  "Roboto",
-  "Roboto Condensed",
-  "Roboto Mono",
-  "The Nautigal",
-  "Titillium Web",
-  "Ubuntu",
-  "Ubuntu Mono",
-].sort();
-
-const fontStyles: string[] = ["bold", "italic", "normal", "bold italic"];
-
-const fontSizes: number[] = [
-  2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 36, 48, 72,
-];
+import { fontFamilies, fontSizes, fontStyles } from "../../../utils/fonts";
 
 function ChangeFontBody({
   fontChangesRef,
